refactor: reorganize exception handling and gym controller structure
