+++ src/main/java/igym/exceptions/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package igym.controller;
+package igym.exceptions;
 
 import java.time.LocalDateTime;
 import java.util.HashMap;
